requirements.txt
streamlit
pandas
numpy
opencv-python-headless
scikit-learn
joblib
tensorflow
pillow